Refactor AdminView and ScheduleMovie component
